+++ TerrainOperations/terrainops.cpp
@@ -1,11 +1,11 @@
 #include "terrainops.h"
 
 namespace terrainops {
-	void to_heightmap(float* data, float* out, int len) {
+	void to_heightmap(float* data, unsigned int* out, int len) {
 
 	}
 
-	void from_heightmap(float* data, float* in, int len) {
+	void from_heightmap(float* data, unsigned int* in, int len) {
 
 	}
 }
